requirements.txt
apipkg==1.4
asn1crypto==0.22.0
astroid==1.6.0
bcrypt==3.1.3
certifi==2017.11.5
cffi==1.11.0
chardet==3.0.4
colorama==0.3.9
coverage==4.4.1
cryptography==2.0.3
cx-Oracle==6.0.3
Django==2.0
django-auth-ldap==1.3.0
django-cors-middleware==1.3.1
django-filter==1.1.0
django-oauth-toolkit==1.0.0
djangorestframework==3.7.7
execnet==1.4.1
Fabric3==1.13.1.post1
gunicorn==19.7.1
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.0.6
paramiko==2.3.0
Pillow==5.0.0
pip==9.0.1
py==1.4.34
pyasn1==0.3.6
pycparser==2.18
pyldap==2.4.45
pylint==1.8.1
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
PyNaCl==1.1.2
pytest==3.2.2
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-forked==0.2
pytest-xdist==1.20.0
pytz==2017.2
requests==2.18.4
setuptools==36.4.0
six==1.11.0
urllib3==1.22
wheel==0.30.0
wrapt==1.10.11
